requirements.txt
# Core dependencies
PyQt6>=6.4.0
hidapi>=0.12.0
pyusb>=1.2.0
libusb-package>=1.2.0

# Advanced features
pywin32>=306
psutil>=5.9.0
pynput>=1.7.6

# Web scraping and requests
requests>=2.28.0
beautifulsoup4>=4.11.0

# Data processing and analytics
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Optional: For enhanced RGB effects
colorsys

# System requirements
# - Windows 10/11 (for full functionality)
# - Python 3.8+
# - Administrative privileges (for USB access)

# Additional dependencies for advanced features
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
